refactor bar
