+++ src/components/line.js
@@ -1,51 +1,8 @@
-import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import ReactEcharts from 'echarts-for-react';
 import _ from 'lodash';
-import { validate } from '../utils';
+import BaseChart from './base';
 
-
-export default class Line extends PureComponent {
-  getSource() {
-    const {
-      source,
-    } = this.props.value;
-    return source;
-  }
-
-  getDimensions() {
-    return _.chain(this.getSource())
-      .first()
-      .keys()
-      .value();
-  }
-
-  getAxisDimension() {
-    return _.first(this.props.value.axisDimensions) ||
-      _.first(this.getDimensions());
-  }
-
-  getAxisData() {
-    const axisDim = this.getAxisDimension();
-    return _.chain(this.getSource())
-      .map(row => row[axisDim])
-      .value();
-  }
-
-  getAxisOption() {
-    return {
-      data: this.getAxisData(),
-      type: 'category',
-      boundaryGap: false,
-    };
-  }
-
-  getMetricDimensions() {
-    return _.isEmpty(this.props.value.metricDimensions) ?
-      _.difference(this.getDimensions(), [this.getAxisDimension()]) :
-      this.props.value.metricDimensions;
-  }
-
+export default class Line extends BaseChart {
   getSeriesOption() {
     const source = this.getSource();
     return _.chain(this.getMetricDimensions())
@@ -74,26 +31,14 @@ export default class Line extends PureComponent {
     });
   }
 
-  render() {
-    const source = this.getSource();
-
-    validate(source);
-    const onEvents = {
+  getEvents() {
+    return {
       click: args =>
         this.props.onSlicerChange(_.defaults(
-          {}, { dataObj: _.zipObject(_.first(source), args.data) },
+          {}, { dataObj: _.zipObject(this.getDimensions(), args.data) },
           args,
         )),
     };
-
-    return (
-      <ReactEcharts
-        option={this.getOption()}
-        notMerge={true} //eslint-disable-line
-        onEvents={onEvents}
-        {...this.props}
-      />
-    );
   }
 }
